+++ src/Ext/WeaponType/Body.cpp
@@ -2,6 +2,9 @@
 #include <Ext/Bullet/Body.h>
 #include <Ext/Techno/Body.h>
 
+#include <BulletTypeClass.h>
+#include <BulletClass.h>
+
 WeaponTypeExt::ExtContainer WeaponTypeExt::ExtMap;
 
 bool WeaponTypeExt::ExtData::HasRequiredAttachedEffects(TechnoClass* pTarget, TechnoClass* pFirer) const